requirements_phase0.txt
# Phase 0 — CPU only, no CUDA required
torch>=2.1.0
transformers>=4.37.0
datasets>=2.16.0
sentence-transformers>=2.3.0
faiss-cpu>=1.7.4
numpy>=1.24.0
